// extensions/google-sheet-activities/src/services/types.ts
export type GetSheetResult = { range: string; majorDimension: string; values: string[][] };

export type Activity = {
  confirmDate: string;
  projectName: string;
  responsibleMail: string;
  date: string;
  time: string;
  company: string;
  switchboard: string;
  name: string;
  title: string;
  email: string;
  cellPhone: string;
  workPhone: string;
  description: string;
  crm2Id: string;
  eventId: string;
  crmId: string;
  crmAccount: string;
  conversationId: string;
  orderWorth: string;
};